fixes bug on add to playlist function, adds it to single movie page

+++ netflix/models.py
@@ -18,10 +18,11 @@ class Playlist(models.Model):
   original_language = models.CharField(max_length=100)
   original_title = models.CharField(max_length=200)
   overview = models.TextField()
-  popularity = models.IntegerField()
+  popularity = models.FloatField()
   poster_path = models.CharField(max_length=200)
   release_date = models.CharField(max_length=200)
-  vote_average = models.IntegerField()
+  vote_average = models.FloatField()
   vote_count = models.IntegerField(default=0)
+
   date_added = models.DateTimeField()
   user_profile = models.ForeignKey(UserProfile, on_delete=CASCADE)
